requirements.txt
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.12.0
python-dotenv==0.19.0
Werkzeug==2.0.1
WTForms==2.3.3
